big rework
